==> code/01_Cleaning_NLP - Karan Mehta.ipynb <==
## Natural Language Processing
## Final Project - Karan Mehta

import pandas as pd
import warnings
warnings.filterwarnings("ignore")

import re
import nltk
from nltk.tokenize import word_tokenize
import matplotlib.pyplot as plt
import multiprocessing

pd.set_option('display.max_rows', 100)
pd.set_option('display.max_columns', None)
pd.set_option('display.max_colwidth', 200)

%%time
df_news_final_project = pd.read_parquet("D:/University of Chicago/Courses/ADSP 32018 IP02 (Winter 2024) Natural Language Processing and Cognitive Computing/Final Project/news_final_project.parquet", engine='pyarrow')
df_news_final_project.shape

df_news_final_project.head(10)

df_news_final_project['date'] = pd.to_datetime(df_news_final_project['date'])
df_news_final_project['year'] = df_news_final_project['date'].dt.year

df_news_final_project['year'].value_counts()

### News Article (Text) Cleaning

num_processors = multiprocessing.cpu_count()
num_processors

workers = num_processors-1

print(f'Using {workers} workers')

%%time
# Removing urls
df_news_final_project['cleaned'] = df_news_final_project['text'].replace(to_replace=r'\b(?:https?://)?(?:www\.)?\w+\.(?:com|net|io|edu|gov|org|co)\b', value='', regex=True)

%%time
# Removing new line character
df_news_final_project['cleaned'] = df_news_final_project['cleaned'].apply(lambda x: re.sub(r'(?:[\n]+)',' ',str(x)))

%%time
# Removing tab
df_news_final_project['cleaned'] = df_news_final_project['cleaned'].apply(lambda x: re.sub(r'(?:[\t]+)',' ',str(x)))

# Removing HTMl tags
df_news_final_project['cleaned'] = df_news_final_project['cleaned'].apply(lambda x: re.sub('<[^>]*>',' ',str(x)))

# Removing CSS tags
df_news_final_project['cleaned'] = df_news_final_project['cleaned'].apply(lambda x: re.sub('<style[^>]*>.*?</style>',' ',str(x)))

# Removing Javascript tags
df_news_final_project['cleaned'] = df_news_final_project['cleaned'].apply(lambda x: re.sub('<script[^>]*>.*?</script>',' ',str(x)))

# Remove all non-ascii characters
df_news_final_project['cleaned'] = df_news_final_project['cleaned'].apply(lambda x: re.sub(r'[^\x00-\x7F]',' ',str(x)))

# Remove all emojis and symbols characters
df_news_final_project['cleaned'] = df_news_final_project['cleaned'].apply(lambda x: re.sub('[\U0001F300-\U0001F5FF\U0001F600-\U0001F64F\U0001F680-\U0001F6FF\U0001F700-\U0001F77F\U0001F780-\U0001F7FF\U0001F800-\U0001F8FF\U0001F900-\U0001F9FF\U0001FA00-\U0001FA6F\U0001FA70-\U0001FAFF\u2600-\u26FF\u2700-\u27BF]',' ',str(x)))

df_news_final_project['cleaned'] = df_news_final_project['cleaned'].apply(lambda x: re.sub('#\w+',' ',str(x)))

### News Article (Title) Cleaning

df_news_final_project['clean_title'] = df_news_final_project['title'].apply(lambda x: re.sub(r'[^\w\s.,]', '', x))

%%time
# Removing 
df_news_final_project['clean_title'] = df_news_final_project['text'].replace(to_replace=r'\b(?:https?://)?(?:www\.)?\w+\.(?:com|net|io|edu|gov|org|co)\b', value='', regex=True)

%%time
# Removing new line character
df_news_final_project['clean_title'] = df_news_final_project['clean_title'].apply(lambda x: re.sub(r'(?:[\n]+)',' ',str(x)))

%%time
# Removing tab
df_news_final_project['clean_title'] = df_news_final_project['clean_title'].apply(lambda x: re.sub(r'(?:[\t]+)',' ',str(x)))

# Removing HTMl tags
df_news_final_project['clean_title'] = df_news_final_project['clean_title'].apply(lambda x: re.sub('<[^>]*>',' ',str(x)))

# Remove all non-ascii characters
df_news_final_project['clean_title'] = df_news_final_project['clean_title'].apply(lambda x: re.sub(r'[^\x00-\x7F]',' ',str(x)))

# Remove all emojis and symbols characters
df_news_final_project['clean_title'] = df_news_final_project['clean_title'].apply(lambda x: re.sub('[\U0001F300-\U0001F5FF\U0001F600-\U0001F64F\U0001F680-\U0001F6FF\U0001F700-\U0001F77F\U0001F780-\U0001F7FF\U0001F800-\U0001F8FF\U0001F900-\U0001F9FF\U0001FA00-\U0001FA6F\U0001FA70-\U0001FAFF\u2600-\u26FF\u2700-\u27BF]',' ',str(x)))

df_news_final_project['clean_title'] = df_news_final_project['clean_title'].apply(lambda x: re.sub('#\w+',' ',str(x)))

df_news_final_project['clean_title'] = df_news_final_project['clean_title'].str.strip()

### Finding Relevant Articles

%%time
# Finding all the relevant articles 
key_tokens = ['artificial intelligence', 'machine learning', 'analytics', 'openai', 'chatgpt', 
              'nlp', 'data', 'llm', 'Autoencoders', 'Predictive model']

pattern = '|'.join(map(re.escape, key_tokens))

df_news_final_project['relevant'] = df_news_final_project['cleaned'].str.contains(pattern, flags=re.IGNORECASE).astype(int)

df_news_final_project['relevant'].value_counts()

df_relevant = df_news_final_project[df_news_final_project['relevant'] == 1]

%%time
# Below words exist at the bottom of most news articles and are not relevant (remnants of the web crawler)
webcrawl_remnants = ['Advertisement', 'Sponsored', 'Copyright', 'Terms of Service', 'Privacy Policy', 'About Us', 
                     'Contact Us', 'Subscribe', 'Share', 'Recommended Reading', 'Follow Us', 'Tag Cloud', 'Archive',
                     'Search', 'Disclaimer', 'Cookies', 'Sitemap', 'Loading', 'Error', 'All rights reserved', 
                     'Newsletter', 'Press Release', 'Banner', 'Login',  'Policy', 'Feedback', 'Join', 'T&C']

pattern = '|'.join(map(re.escape, webcrawl_remnants))

df_relevant['cleaned'] = df_relevant['cleaned'].str.replace(pattern, ' ', regex=True, flags=re.IGNORECASE)

df_relevant['cleaned'] = df_relevant['cleaned'].str.strip()

%%time
df_relevant['tokens'] = df_relevant['cleaned'].apply(lambda x: word_tokenize(x))

#%%time
#df_relevant['title_token'] = df_relevant['clean_title'].apply(lambda x: word_tokenize(x))

df_relevant['no_tokens'] = df_relevant['tokens'].apply(len)

df_relevant.head(5)

### Saving the Cleaned Data


df_relevant.to_parquet('CleanedDataset.parquet')
