==> code/05_NLP_NER_NewsText_FinalProject.ipynb <==
import pandas as pd

pd.set_option('display.max_rows', 100)
pd.set_option('display.max_columns', None)
pd.set_option('display.max_colwidth', 500)

import nltk
import numpy as np
import re
import string

import seaborn as sns
import matplotlib.pyplot as plt
from nltk.chunk import ne_chunk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import word_tokenize
from collections import Counter
import warnings
warnings.filterwarnings("ignore")

from pandarallel import pandarallel
import spacy

from google.cloud import storage
# Instantiates a client
storage_client = storage.Client()

# The name for the new bucket
bucket_name = "nlp_finalproject"

# Creates the new bucket
bucket = storage_client.bucket(bucket_name)
print(f"Bucket {bucket.name} connected.")

df_final5 = pd.read_parquet("gs://nlp_finalproject/FilteredDataset.parquet", engine='pyarrow')
df_final5.head(5)

# !pip install -U pip setuptools wheel
# !pip install -U spacy
# !python -m spacy download en_core_web_sm

!python -m spacy download en_core_web_sm

### Cleaning to find ORG's in News Text

pandarallel.initialize(progress_bar=True, use_memory_fs=False)

df_final5 = df_final5.drop(['relevant','tokens','url','language'],axis=1)

df_final5['NER_clean'] = df_final5['cleaned'].apply(lambda x: re.sub(r'[^\w\s.,]', '', x))

df_final5['NER_clean'] = df_final5['NER_clean'].str.replace('\d+', '', regex=True)

df_final5['NER_clean'] = df_final5['NER_clean'].apply(lambda x: re.sub(r'{So}', '', x))

df_final5['NER_clean'] = df_final5['NER_clean'].str.strip()

df_final5['year'] = df_final5['date'].dt.year

df_final5['year'].value_counts()

### NER to Find ORG and Person in News text

nlp = spacy.load("en_core_web_sm")

df_final5 = df_final5.drop(['no_tokens','percentage'],axis=1)

def extract_entities(text):
    doc = nlp(text)
    entities = [ent.text for ent in doc.ents if ent.label_ == "ORG"]
    return entities

df_final5['ORG_Ent'] = df_final5['NER_clean'].parallel_apply(extract_entities)

%%time
def extract_entities(text):
    doc = nlp(text)
    entities = [ent.text for ent in doc.ents if ent.label_ == "Person"]
    return entities

df_final5['Person_Ent'] = df_final5['NER_clean'].parallel_apply(extract_entities)

df_final5 = df_final5.drop('NER_clean',axis=1)

df_final5.head(5)

### Saving the NER Dataframe

df_final5.to_parquet('gs://nlp_finalproject/NER_Clean.parquet')
