==> code/07_Sentiment Analysis - NLP.ipynb <==
import pandas as pd

import nltk
import numpy as np
import re
import string

import seaborn as sns
import matplotlib.pyplot as plt
from nltk.chunk import ne_chunk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import word_tokenize
from collections import Counter
import warnings
warnings.filterwarnings("ignore")

import spacy
!python -m spacy download en_core_web_sm

import multiprocessing
from pandarallel import pandarallel
import spacy
from nltk.sentiment.vader import SentimentIntensityAnalyzer

!pip install vaderSentiment

from google.cloud import storage
storage_client = storage.Client()

bucket_name = "nlp_finalproject"

bucket = storage_client.bucket(bucket_name)
print(f"Bucket {bucket.name} connected.")

df_final_sent = pd.read_parquet("gs://nlp_finalproject/NER_Clean.parquet", engine='pyarrow')
df_final_sent.head(5)

nltk.download('punkt')
nltk.download('stopwords')
nltk.download('vader_lexicon')

num_processors = multiprocessing.cpu_count()
pandarallel.initialize(nb_workers=num_processors-1, use_memory_fs=False, progress_bar = True)

from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
sid = SentimentIntensityAnalyzer()

def analyze_sentiment(text):
    sentiment_scores = sid.polarity_scores(text)
    return sentiment_scores['compound']

df_final_sent['sentiment_score'] = df_final_sent['cleaned'].parallel_apply(analyze_sentiment)

df_final_sent['sentiment'] = df_final_sent['sentiment_score'].apply(lambda score: 'Positive' if score > 0.95 else ('Negative' if score < -0.90 else 'Neutral'))

df_final_sent['sentiment'].value_counts()

# Displaying the data with sentiment labels
df_final_sent[['title','cleaned', 'sentiment','sentiment_score']]

df_final_sent = df_final_sent.drop(['Person_Ent'],axis=1)

df_final_sent.head(5)

nlp = spacy.load("en_core_web_sm")

%%time
def extract_entities(text):
    doc = nlp(text)
    entities = [ent.text for ent in doc.ents if ent.label_ == "PERSON"]
    return entities

df_final_sent['Person_Ent'] = df_final_sent['cleaned'].parallel_apply(extract_entities)

df_final_sent[['Person_Ent','year']]

tmp = df_final_sent[['Person_Ent','year']]
plot_tmp = tmp.explode('Person_Ent')
Person_counts = plot_tmp['Person_Ent'].value_counts()

Person_counts.head(30)

df_final_sent.to_parquet('gs://nlp_finalproject/SentimentAnalysis_Clean.parquet')
