==> code/08_Sentiment Analysis (Yelp) -  NLP.ipynb <==
import pandas as pd

import nltk
import numpy as np
import re
import string

import seaborn as sns
import matplotlib.pyplot as plt
from collections import Counter
import warnings
warnings.filterwarnings("ignore")

import multiprocessing
from pandarallel import pandarallel
import spacy

from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.linear_model import SGDClassifier
from sklearn.pipeline import make_pipeline
from sklearn.metrics import classification_report, confusion_matrix, accuracy_score
from sklearn.linear_model import LogisticRegression
from sklearn import metrics

import nltk
from nltk.tokenize import sent_tokenize
nltk.download('punkt')

from google.cloud import storage
storage_client = storage.Client()

bucket_name = "nlp_finalproject"

bucket = storage_client.bucket(bucket_name)
print(f"Bucket {bucket.name} connected.")

df_yelp_sent = pd.read_parquet("gs://nlp_finalproject/SentimentAnalysis_Clean.parquet", engine='pyarrow')
df_yelp_sent.head(5)

num_processors = multiprocessing.cpu_count()
pandarallel.initialize(nb_workers=num_processors-1, use_memory_fs=False, progress_bar = True)

directory = 'https://storage.googleapis.com/msca-bdp-data-open/yelp/'
fileName = 'yelp_train_sentiment.json'

path = directory + fileName

yelp = pd.read_json(path, orient='records', lines=True)

X = yelp['text']
y = yelp['label']

# split X and y into training and testing sets
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=1)

pipe_svm = make_pipeline(
    CountVectorizer(lowercase=False, stop_words='english', ngram_range=(1,3)),
    SGDClassifier(max_iter=100, tol=None,loss='log_loss') 
)

pipe_svm.fit(X_train, y_train)

y_pred = pipe_svm.predict(X_test)

probabilities = pipe_svm.predict_proba(X_test)

def sentiment_score(text):
    prob = pipe_svm.predict_proba([text]) 
    return prob[0]

df_yelp_sent['yelp_sent_score'] = df_yelp_sent['cleaned'].parallel_apply(sentiment_score)

def sentiment(probabilities):
    if probabilities[0] >= 0.85:
        return 'Negative'
    elif probabilities[1] >= 0.85:
        return 'Positive'
    else:
        return 'Neutral'
        
df_yelp_sent['sentiment_yelp'] = df_yelp_sent['yelp_sent_score'].parallel_apply(sentiment)

df_yelp_sent['sentiment_yelp'].value_counts()

!python -m spacy download en_core_web_sm

nlp = spacy.load("en_core_web_sm")

def extract_entities(text):
    doc = nlp(text)
    entities = [ent.text for ent in doc.ents if ent.label_ == "GPE"]
    return entities

df_yelp_sent['GPE_Ent'] = df_yelp_sent['cleaned'].parallel_apply(extract_entities)

def extract_entities(text):
    doc = nlp(text)
    entities = [ent.text for ent in doc.ents if ent.label_ == "PRODUCT"]
    return entities

df_yelp_sent['PROD_Ent'] = df_yelp_sent['cleaned'].parallel_apply(extract_entities)

df_yelp_sent.to_parquet('gs://nlp_finalproject/FinalData.parquet')

### Result for Yelp Sentiment Analysis Using SVM

- Negative    94440
- Positive    40630
- Neutral     28251

### Result for Yelp Sentiment Analysis Using Logistic Regression

- Negative    103568
- Positive     45077
- Neutral      14676
